+++ src/influx_db.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timezone
-from typing import Any, Dict, Set, Type, Union
+from typing import Any, ClassVar, Dict, Set, Type, Union
 
 from pydantic import BaseModel, Field
 
@@ -27,6 +27,8 @@ class InfluxDBLine(BaseModel, metaclass=FieldAccessMeta):
     _VALID_TYPES: Set[Type] = {str, int, bool, float}
     _VALID_TYPE_NAMES: Dict[str, Type] = {_.__name__: _ for _ in _VALID_TYPES}
 
+    _merge_on: ClassVar[str] = ""
+
     @classmethod
     def _parse_value(cls, value_type: str, value: str) -> _Scalar:
         if value_type not in cls._VALID_TYPE_NAMES:
@@ -53,9 +55,14 @@ class InfluxDBLine(BaseModel, metaclass=FieldAccessMeta):
             tags=tags,
         )
 
+    @classmethod
+    def merge_data_points_on(cls, new_pattern: str) -> None:
+        """Merge instances with identical string formatted using ``self.__dict__.`."""
+        cls._merge_on = new_pattern
+
     @property
     def merge_id(self) -> str:
-        return f"{self.measurement}/{self.bucket}/{self.tags}"
+        return self._merge_on.format(**self.__dict__)
 
     def merge(self, other: "InfluxDBLine") -> "InfluxDBLine":
         if not isinstance(other, InfluxDBLine):